+++ src/app/components/customer/customer.component.ts
@@ -3,6 +3,10 @@ import { MatTableDataSource, MatSort, MatPaginator } from '@angular/material';
 import { Router } from '@angular/router';
 import { CustomerService } from 'src/app/_services/customer.service';
 import { Customer } from 'src/app/_models/customer';
+import { switchMap } from 'rxjs/operators';
+import { HttpErrorResponse } from '@angular/common/http';
+import { MessagesService } from 'src/app/_services/message.service';
+import { ConfirmService } from 'src/app/_services/confirm.service';
 
 @Component({
   selector: 'app-customer',
@@ -17,7 +21,9 @@ export class CustomerComponent implements OnInit {
   searchKey: string;
 
   constructor(private router: Router,
-    private customerService: CustomerService) { }
+    private customerService: CustomerService,
+    private messagesService: MessagesService,
+    private confirmService: ConfirmService) { }
 
   ngOnInit() {
     this.customerService.getCustomers().then(customer => {
@@ -42,10 +48,16 @@ export class CustomerComponent implements OnInit {
   }
 
   deleteCustomer(customer: Customer): void {
-    this.customerService.deleteCustomer(customer)
-      .then(data => {
-        console.log(data);
-        this.customerService.getCustomers().then(customer => {
+    const name = 'Delete ' + customer.name + '?';
+    this.confirmService.confirm(name, 'This action is final. Gone forever!').pipe(
+      switchMap(res => {if (res === true) {
+        return  this.customerService.deleteCustomer(customer);
+		}}))
+      .subscribe(
+        result => {
+          this.success();
+          // Refresh DataTable to remove row.
+          this.customerService.getCustomers().then(customer => {
           this.listData = new MatTableDataSource(customer);
           this.listData.sort = this.sort;
           this.listData.paginator = this.paginator;
@@ -55,10 +67,19 @@ export class CustomerComponent implements OnInit {
             });
           };
         });
-      });
+		 
+        },
+        (err: HttpErrorResponse) => {
+          this.messagesService.openDialog('Error', 'Delete did not happen.');
+        }
+      );
   };
 
   edit(customer: Customer) {
     this.router.navigate(['/customer/edit',customer.id]);
+  }
+
+  private success() {
+    this.messagesService.openDialog('Success', 'Database updated as you wished!');
   }
 }
