+++ src/app/components/department/department.component.ts
@@ -3,6 +3,10 @@ import { MatTableDataSource, MatSort, MatPaginator } from '@angular/material';
 import { DepartmentService } from 'src/app/_services/department.service';
 import { Router } from '@angular/router';
 import { Department } from 'src/app/_models/department';
+import { HttpErrorResponse } from '@angular/common/http';
+import { switchMap } from 'rxjs/operators';
+import { ConfirmService } from 'src/app/_services/confirm.service';
+import { MessagesService } from 'src/app/_services/message.service';
 
 @Component({
   selector: 'app-department',
@@ -17,7 +21,9 @@ export class DepartmentComponent implements OnInit {
   searchKey: string;
 
   constructor(private router: Router,
-    private departmentService: DepartmentService) { }
+    private departmentService: DepartmentService,
+    private messagesService: MessagesService,
+    private confirmService: ConfirmService) { }
 
   ngOnInit() {
     this.departmentService.getDepartments().then(dept => {
@@ -42,10 +48,16 @@ export class DepartmentComponent implements OnInit {
   }
 
   deleteUser(department: Department): void {
-    this.departmentService.deleteDepartment(department)
-      .then(data => {
-        console.log(data);
-        this.departmentService.getDepartments().then(dept => {
+    const name = 'Delete ' + department.name + '?';
+    this.confirmService.confirm(name, 'This action is final. Gone forever!').pipe(
+      switchMap(res => {if (res === true) {
+        return this.departmentService.deleteDepartment(department)
+		}}))
+      .subscribe(
+        result => {
+          this.success();
+          // Refresh DataTable to remove row.
+          this.departmentService.getDepartments().then(dept => {
           this.listData = new MatTableDataSource(dept);
           this.listData.sort = this.sort;
           this.listData.paginator = this.paginator;
@@ -55,11 +67,18 @@ export class DepartmentComponent implements OnInit {
             });
           };
         });
-      });
+        },
+        (err: HttpErrorResponse) => {
+          this.messagesService.openDialog('Error', 'Delete did not happen.');
+        }
+      );
   };
 
   edit(department: Department) {
     this.router.navigate(['/department/edit',department.id]);
   }
 
+  private success() {
+    this.messagesService.openDialog('Success', 'Database updated as you wished!');
+  }
 }
